TestDataUtils cleanup
